+++ integracaoAPIs/models/__pycache__/implement.cpython-310.pyc
