+++ integracaoAPIs/models/__pycache__/integracaoBD.cpython-310.pyc
