+++ integracaoAPIs/models/__pycache__/models.cpython-310.pyc
