+++ integracaoAPIs/models/__pycache__/operacoesBD.cpython-310.pyc
