+++ integracaoAPIs/models/__pycache__/services.cpython-310.pyc
